change time-variable from String to Tuple

+++ HasKAL/src/HasKAL/ExternalUtils/LAL/Lwtprint.hs
@@ -1,7 +1,7 @@
 {-******************************************
   *     File Name: Lwtprint.hs
   *        Author: Takahiro Yamamoto
-  * Last Modified: 2014/06/17 19:34:04
+  * Last Modified: 2014/06/18 20:54:32
   *******************************************-}
 
 module HasKAL.ExternalUtils.LAL.Lwtprint
@@ -12,10 +12,10 @@ module HasKAL.ExternalUtils.LAL.Lwtprint
 import qualified System.Process as SP -- readProcess
 import qualified Control.Monad as CM -- forM
 
-execLwtprint :: [String] -> String -> Int -> Int -> [String] -> IO [String]
+execLwtprint :: [String] -> String -> Integer -> Int -> [String] -> IO [String]
 execLwtprint glitchActiveLabels kwBasename kwGpsTime kwStride kwActiveLabels = do
    CM.forM [0..(length glitchActiveLabels)-1] $ \lambda -> do
-      let tempGps = kwGpsTime + kwStride 
+      let tempGps = kwGpsTime + (fromIntegral kwStride)
       let filePath = kwBasename ++ (glitchActiveLabels !! lambda) ++ "-" ++ (take 5 $ show kwGpsTime) ++ "/" ++ kwBasename ++ (glitchActiveLabels !! lambda) ++ "-" ++ (show tempGps) ++ "-" ++ (show kwStride) ++ ".xml"
       let cmd_string = [ filePath ,("-c " ++ (unwords kwActiveLabels)) ]
       SP.readProcess "lwtprint" cmd_string [] -- execute lwtprint
